When setting a null device for an audio output, force the next routing request.
When a null device is set for an audio ouput, the outputDesc->mDevice is not updated,
thus the routing update is not done next time the device is selected for that output.
To meet the device change conditions, a boolean member mForceRouting has been
introduced to force the routing next time the output is set with a non-null device.

Change-Id: I5571ed55dadce2b9751c9e2101a5d1cfca3e8479
Origin-Change-Id: I370126e4175cc797b684945cba97f4b92c5eb8a1
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Jian Luo <jian.luo@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 53856