AudioPolicy: improve stopInput
Given the convention for audio device identification since JB MR1 code
drop, the bit 31 of the bit field is reserved and when set indicates
an input device. So 0x0 (AUDIO_DEVICE_NONE) is identified as an output
audio device. Thus the stopInput function does not stop the audio input
stream.

In order to reroute the audio input stream as expected, this patch
modifies the stopInput function in order to use the
AUDIO_DEVICE_BIT_IN (0x80000000) value.

Change-Id: I05fe4e3ac98cae2eb2cde951abd284b8b6540dbe
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>
Signed-off-by: Jian Luo <jian.luo@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 104921